Use size_t in allocator & allocate at least 16 bytes & update defragmentation logic.

+++ runtime/memory_resource/memory_resource.cpp
@@ -4,11 +4,11 @@
 
 namespace {
 
-void write_stat(stats_t *stats, const char *prefix, const char *suffix, long long int value) noexcept {
+void write_stat(stats_t *stats, const char *prefix, const char *suffix, size_t value) noexcept {
   char buffer[256]{0};
   const int len = snprintf(buffer, sizeof(buffer) - 1, "%s.%s", prefix, suffix);
   assert(len > 0 && sizeof(buffer) >= static_cast<size_t>(len + 1));
-  add_histogram_stat_long(stats, buffer, value);
+  add_histogram_stat_long(stats, buffer, static_cast<int64_t>(value));
 };
 
 } // namespace
